+++ Spreadsheet_Manjesh_Puram/SpreadsheetEngine/MultiplyOperatorNode.cs
@@ -26,7 +26,7 @@ namespace CptS321
         /// <summary>
         /// Gets, this contains the information needed to know what precedence the multiplication has.
         /// </summary>
-        public ushort Precedence { get; } = 1;
+        public int Precedence { get; } = 1;
 
         /// <summary>
         /// This function serves the purpose to return the multiplication when we do the evaluation for the calculations.
